% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fredundancy.R
\name{fredundancy}
\alias{fredundancy}
\title{Compute measures of functional redundancy and functional interdependency based on relative entropy}
\usage{
fredundancy(functions, abundance , n_reference = NULL)
}
\arguments{

\item{functions}{The paired vector of functions for the abundance vector. Needs to be the same size as the abundance vector.
The i-th component of \code{functions} corresponds to the i-th abundance of the species, encoded in the abundance vector.
Gets transformed into relative frequencies, such that the components sum up to 1 and its components correspond to relative frequencies. Might contain zeros.}

\item{abundance}{A vector of non-zero abundances representing each species in the community. The vector is checked to ensure the abundances sum to 1.
              If they do not, the vector is normalized, provided the sum is within an acceptable tolerance (*|1 - sum(abundances)| < 1e-5*). If the sum falls outside this tolerance, an error message is returned.}

\item{n_reference}{Optional value to compute reference-based redundancy. It is an integer value of the number of species in the reference
that can perform the function.}
}
\value{
A list containing:
\itemize{
\item \code{sample_based}: A coefficient corresponding to taxon sample-based redundancy.
\item \code{reference_based}: A coefficient corresponding to taxon reference-based redundancy (optional, if \code{n_reference} is set).
\item \code{abundance_based}: A coefficient corresponding to abundance-based redundancy.
\item \code{interdependency}: Interdependency measure.
}
}
\description{
Compute measures of functional redundancy and functional interdependency based on relative entropy
}
\examples{
abundance <- c(0.2, 0.1, 0.05, 0.05, 0.6)
functions <- c(0.8, 0.1, 0.05, 0.05, 0)
n_reference <- 8
fredundancy(functions = functions, abundance = abundance, n_reference = NULL)

}
